Move Android Host Structures into AndroidStructures Header
