+++ menrva/app/src/main/cpp/command_handlers/Module_GetConfig_Handler.cpp
@@ -17,7 +17,7 @@
  */
 
 #include "Module_GetConfig_Handler.h"
-#include "../hosts/android/AndroidInterface.h"
+#include "../hosts/android/AndroidStructures.h"
 
 Module_GetConfig_Handler::Module_GetConfig_Handler(LoggerBase* logger)
         : TypedCommandHandlerBase(new Android_SystemCommand_Command(), logger, __PRETTY_FUNCTION__) {
