+++ menrva/app/src/main/cpp/command_handlers/Module_SetConfig_Handler.cpp
@@ -18,7 +18,7 @@
 
 #include "Module_SetConfig_Handler.h"
 #include "../tools/StringOperations.h"
-#include "../hosts/android/AndroidInterface.h"
+#include "../hosts/android/AndroidStructures.h"
 
 Module_SetConfig_Handler::Module_SetConfig_Handler(LoggerBase* logger)
         : TypedCommandHandlerBase(new Android_SystemCommand_Command(), logger, __PRETTY_FUNCTION__) {
